+++ R/zzz.R
@@ -11,5 +11,7 @@
   register_rule(rule_antipattern_setwd())
   register_rule(rule_antipattern_hardcoded_path())
   register_rule(rule_documentation_missing())
+  register_rule(rule_testing_missing_tests())
+  register_rule(rule_package_deprecated_api())
   invisible()
 }